requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-crispy-forms==1.12.0
environs==9.3.2
greenlet==1.1.0
idna==2.10
Jinja2==2.11.3
Pillow==8.3.1
psutil==5.8.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
SQLAlchemy==1.4.11
sqlparse==0.4.1
djangorestframework==3.12.4
beautifulsoup4==4.9.3
bs4==0.0.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
geopy~=2.2.0
environs~=9.3.2